requirements.txt
flask
flask_cors
boto3
python-dotenv
stripe